Make comments ADDED/REMOVED/CHANGED
Have three different types of comments:

 Patch.ADDED: Comments on new lines
 Patch.REMOVED: Comments on removed lines
 Patch.CHANGED: Comments on changed segments

ADDED/REMOVED are represented in the review format by ommitting
the - or the + of @@ -I,N +J,M @@ hunk intro.

'... N more ...' abbrevations are introduced into the review
format to allow representing a comment on an entire changed
segment without getting really long.

The CSS and HTML is adjusted so that the comments appear in
three overlapping vertical columns based on their type; this is
done by adding a single column-spanned columnArea div under
the line and then using floats within it.

Patch.File.getLocation() now allows either oldLine or newLine
to be null to look up by only one or the other.