Add review comment summary to the Overview page
Instead of just showing comments inline, show them on the overview
page as well with context formatted similarly to the two-column
patch display.

Both old comments, and draft comments for the review are shown;
click on an old comment jumps to that comment. Clicking on a
draft comment jumps to it and starts editing it again.

Move "Your Review" above "Previous Reviews" so that you have easy
access to editing it when switching pages.