Create new extension - TemplateOverrides
A central place where template overrides are kept and their original
counterparts' sha256sums are checked.

This is to make sure that we are overriding a known version of
template. If sha256sum does not agree then it means that the original
has changed and there may be some fixes to be backported to our
overrides.