Automatically link Git commits to the corresponding URL
Automatically link "commit 1a2b3c4d" and "pushed as 1a2b3c4d" to
the corresponding URL for that commit on https://git.gnome.org.

The ID can have between 8 and 40 characters. The URL construction
relies on the product name of the Bugzilla report so if the name
in Bugzilla does not match the repository name in Git the URL
will be a 404. So you might want to make sure that names are the
same in Bugzilla and Git if you want this feature to work.

Patch written by Olav Vitters and Andre Klapper.

Bug 559537