Automatically link Git wip/ branches to the corresponding URL
Automatically link "branch wip/$branch" to the corresponding URL for that branch
on https://git.gnome.org. The URL construction relies on the product name of the
Bugzilla report so if the name in Bugzilla does not match the repository name in
Git the URL will be a 404. So you might want to make sure that names are the
same in Bugzilla