[Browse] Remove NEEDINFO section; just put a single query in sidebar
Apart from triagers as a target group there is no real need to
differentiate on five levels of age for tickets in NEEDINFO status.
The previous visibility and space used by that was not justified.

Put one single query in the sidebar for any NEEDINFO tickets in
that product, sort the search results of that query by oldest
tickets first, and force displaying the Last Changed column on
those results so it becomes clear how old tickets are.