Move Priority and Severity sections from sidebar to main area
...and put them next to the Target Milestones. As priority and
severity have a limited number of choices and as maintainers
should clean up those old open tasks with past target milestones
this looks like a good combination when it comes to the height of
these sections.

Furthermore, proper usage of priority and severity in combination
with target milestones seems like important enough to not hide
this in the sidebar but make it more central.